Started to validate the login form.
Also fixed references to the new components and incorporated
the new application patterns.

Refs: #111
