--- firenado/examples/skell/components/internal/__init__.py
@@ -1,27 +1,0 @@
-#!/usr/bin/env python
-#
-# Copyright 2015-2016 Flavio Garcia
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import firenado.tornadoweb
-from skell.components.internal import uimodules
-
-
-class SkellInternalComponent(firenado.tornadoweb.TornadoComponent):
-    """ This component just adds another ui modules package to the application
-    ui modules stack.
-    """
-
-    def get_ui_modules(self):
-        return uimodules
