+++ firenado/examples/skell/templates/login.html
@@ -1,0 +1,14 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <form class="form-signin" action="login" method="post">
+        <h2 class="form-signin-heading">Please sign in</h2>
+        <label for="username" class="sr-only">User Name</label>
+        <input type="login" id="username" name="username" class="form-control" placeholder="User Name" autofocus>
+        {% module LoginErrorMessage("username") %}
+        <label for="password" class="sr-only">Password</label>
+        <input type="password" id="password" name="password" class="form-control" placeholder="Password">
+        {% module LoginErrorMessage("password") %}
+        <button class="btn btn-lg btn-primary btn-block" type="submit">Sign in</button>
+    </form>
+{% end %}
