+++ firenado/examples/skell/templates/uimodules/login_error_message.html
@@ -1,0 +1,5 @@
+<div class="alert alert-danger" role="alert">
+  <span class="glyphicon glyphicon-exclamation-sign" aria-hidden="true"></span>
+  <span class="sr-only">Error:</span>
+  {{ message }}
+</div>
